Added caching mechanism:
Explanation: "Piwi supports caching. That means you can specify a timespan after which pages are regenerated and not read from cache.
Piwi is differing between different parameters, so calling the same page with different parameters will result in multiple cached pages.

To configure caching, just edit the settings within your 'site.xml'. Set the attribute 'cachetime' to the desired number of seconds. To disable caching set it to 0 or delete the attribute.

Currently it is only possible to define one global 'cachetime', for the future it is planed to also allow specific settings for each page."